Move the logic responsible for creating the initial store from FamilyGroupService
