+++ composeApp/src/commonMain/kotlin/com/github/familyvault/services/FileCabinetService.kt
@@ -3,6 +3,8 @@ package com.github.familyvault.services
 import com.github.familyvault.AppConfig
 import com.github.familyvault.backend.client.IPrivMxClient
 import com.github.familyvault.backend.models.ThreadItem
+import com.github.familyvault.models.enums.StoreType
+import com.github.familyvault.models.enums.fileCabinet.FileCabinetThreadType
 import com.github.familyvault.utils.FamilyMembersSplitter
 
 class FileCabinetService(
@@ -11,6 +13,34 @@ class FileCabinetService(
     private val familyGroupSessionService: IFamilyGroupSessionService,
     private val familyGroupService: IFamilyGroupService
 ) : IFileCabinetService {
+    override suspend fun createInitialStores() {
+        val contextId = familyGroupSessionService.getContextId()
+        val familyGroupName = familyGroupSessionService.getFamilyGroupName()
+        val splitFamilyMembers = FamilyMembersSplitter.split(
+            familyGroupService.retrieveFamilyGroupMembersList()
+        )
+
+        val users = splitFamilyMembers.members.map { it.toPrivMxUser() }
+        val managers = splitFamilyMembers.guardians.map { it.toPrivMxUser() }
+
+        val storeId = privMxClient.createStore(
+            contextId,
+            users,
+            managers,
+            StoreType.FILE_CABINET_IMAGES.toString()
+        )
+
+        privMxClient.createThread(
+            contextId = contextId,
+            users = users,
+            managers = managers,
+            tag = AppConfig.FILE_CABINET_THREAD_TAG,
+            type = FileCabinetThreadType.IMAGES.toString(),
+            name = familyGroupName,
+            referenceStoreId = storeId,
+            threadInitialCreators = emptyList()
+        )
+    }
 
     override fun retrieveFileCabinetThread(): ThreadItem {
         val contextId = familyGroupSessionService.getContextId()
