+++ composeApp/src/commonMain/kotlin/com/github/familyvault/services/IFileCabinetService.kt
@@ -3,6 +3,7 @@ package com.github.familyvault.services
 import com.github.familyvault.backend.models.ThreadItem
 
 interface IFileCabinetService {
+    suspend fun createInitialStores()
     fun retrieveFileCabinetThread(): ThreadItem
     fun retrieveFileCabinetStoreId(): String
